# src/rest-server/docs/swagger.yaml
openapi: 3.0.1
info:
  title: OpenPAI RESTful API
  description: >
    Open Platform for AI RESTful API docs.
    Version 2.0.1: add more examples and fix typos
    Version 2.0.2: update job detail and job attempt schema
    Version 2.0.3: update parameters description of get storage list, update storage example and add get job config example
    Version 2.0.4: add default field in get storage list
    Version 2.0.5: add more parameters to job list; add submissionTime
    Version 2.1.0: add add/delete tag api; add tags field in get job detail and get job list; add tags filter in get job list
    Version 2.1.1: add get event api
    Version 2.2.0: add get task status api; add jobAttempId filter to job status api and extend job detail schema
    Version 2.2.1: a user can add/delete tags to/from his/her own jobs
    version 2.2.2: add get task logs api
    version 2.2.3: set .jobStatus.appId nullable in get jobAttempt schema
    version 2.2.4: support sorting by completionTime in get the list of jobs
    version 2.2.5: add alert related api
    version 2.2.6: update type of taskUid
    version 2.2.7: add jobPriority list jobs parameter
  license:
    name: MIT License
    url: "https://github.com/microsoft/pai/blob/master/LICENSE"
  version: 2.2.7
...
paths:
...
  /api/v2/tokens:
    get:
      tags:
        - token
      summary: Get your currently signed tokens.
      description: Get your currently signed tokens.
      operationId: getTokens
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      type: string
                    description: JWT string list
                required:
                  - tokens
              example:
                tokens:
                  - "JWT Token Example"
        "401":
          $ref: "#/components/responses/UnauthorizedUserError"
    delete:
      tags:
        - token
      summary: Revoke all tokens.
      description: Revoke all tokens.
      operationId: deleteTokens
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                message: revoke successfully
        "401":
          $ref: "#/components/responses/UnauthorizedUserError"
        "403":
          $ref: "#/components/responses/ForbiddenUserError"
  "/api/v2/tokens/{token}":
    delete:
      tags:
        - token
      summary: Revoke a token.
      description: Revoke a token.
      operationId: deleteToken
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/token"
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
              example:
                message: revoke successfully
        "401":
          $ref: "#/components/responses/UnauthorizedUserError"
        "403":
          $ref: "#/components/responses/ForbiddenUserError"
  /api/v2/tokens/application:
    post:
      tags:
        - token
      summary: Create an application access token in the system.
      description: >
        Application access token can only be used for job related operations.
        Application access token has no expiration time and can be revoked
        manually.
      operationId: createApplicationToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: your access token
                  application:
                    type: boolean
                    description: "true"
              example:
                token: "JWT Token Example"
                application: true
        "401":
          $ref: "#/components/responses/UnauthorizedUserError"
...
components:
  parameters:
    token:
      name: token
      in: path
      description: token
      required: true
      schema:
        type: string
      example: test-token
...
  schemas:
    Response:
      type: object
      properties:
        code:
          type: string
          enum:
            - BadConfigurationError
            - ConflictJobError
            - ConflictUserError
            - ConflictVcError
            - ForbiddenUserError
            - ForbiddenKeyError
            - IncorrectPasswordError
            - InvalidParametersError
            - NoApiError
            - NoJobError
            - NoJobConfigError
            - NoJobSshInfoError
            - NoUserError
            - NoGroupError
            - NoVirtualClusterError
            - ReadOnlyJobError
            - RemoveAdminError
            - ReadOnlyVcError
            - RemoveRunningVcError
            - UnauthorizedUserError
            - NoEnoughQuotaError
            - NotImplementedError
            - UnknownError
        message:
          type: string
      required:
        - message
...
  responses:
...
    UnauthorizedUserError:
      description: UnauthorizedUserError
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
          examples:
            UnauthorizedUserError:
              value:
                code: UnauthorizedUserError
                message: Guest is not allowed to do this operation.
    ForbiddenUserError:
      description: ForbiddenUserError
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response"
          examples:
            ForbiddenUserError:
              value:
                code: ForbiddenUserError
                message: Non-admin is not allow to do this operation.
            ForbiddenTokenError:
              value:
                code: ForbiddenUserError
                message: Applications are not allowed to do this operation.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT